Added initial DICOM medical imagery importer.
This importer should open most imagery from DICOM files and copy most of the
metadata into the Opticks metadata. Currently, JPEG2000 will not load
and a few other obscure bit formats may have problems. Also, the metadata is
not interpreted except to convert to the appropriate base type (int, etc.)
when appropriate. There's still quite a bit of work to do as far as metadata
interpretation and testing. This is a good start. JPEG2000 can be supported
with a non-free extension to dcmtk. Maybe we should write an open source
extension for JPEG2000?